Added version of LeafAligner without map

+++ aligner.hpp
@@ -14,7 +14,7 @@ class Result {
 
     Result(unsigned int read_start, 
            unsigned int seed_len, 
-           double prob, 
+           float prob, 
            unsigned int ref_start = 0, 
            unsigned int ref_end = 0);
 
@@ -22,10 +22,10 @@ class Result {
     void print(std::ostream &out);
 
     Range read_range_, ref_range_;
-    double seed_prob_;
+    float seed_prob_;
 
     #ifdef DEBUG_PROB
-    double min_evt_prob_;
+    float min_evt_prob_;
     #endif
 };
 
@@ -38,32 +38,32 @@ class AlnParams {
                  max_rep_copy_,
                  max_paths_;
 
-    double max_stay_frac_;
+    float max_stay_frac_;
 
     unsigned int max_consec_stay_, 
                  max_ignores_,
                  max_skips_;
 
-    double window_prob_;
+    float window_prob_;
 
     std::vector<unsigned int> evpr_lengths_;
-    std::vector<double> evpr_threshes_;
+    std::vector<float> evpr_threshes_;
 
     AlnParams(const KmerModel &model,
               unsigned int path_win_len, 
               unsigned int min_rep_len, 
               unsigned int max_rep_copy, 
               unsigned int max_paths, 
-              double max_stay_frac,
+              float max_stay_frac,
               unsigned int max_consec_stay,
               unsigned int max_ignores, 
               unsigned int max_skips,
               const std::string event_probs,
-              double window_prob);           
+              float window_prob);           
     
     unsigned int nucl_to_events(unsigned int n);
-    double get_prob_thresh(unsigned int fm_length);
-    double get_source_prob();
+    float get_prob_thresh(unsigned int fm_length);
+    float get_source_prob();
 };
 
 
@@ -73,7 +73,7 @@ class Aligner {
 
     virtual void new_read(size_t read_len) = 0;
     virtual void reset() = 0;
-    virtual std::vector<Result> add_event(double *kmer_probs, std::ostream &seeds_out, std::ostream &time_out) = 0;
+    virtual std::vector<Result> add_event(float *kmer_probs, std::ostream &seeds_out, std::ostream &time_out) = 0;
 };
 
 #endif
